+++ packages/core/src/index.ts
@@ -9,6 +9,7 @@ export * from "./scanners/rls-sql.js";
 export * from "./scanners/storage-acl.js";
 export * from "./scanners/nextjs-routes.js";
 export * from "./scanners/firebase-rules.js";
+export * from "./scanners/client-env.js";
 export * from "./remediation/secret-pr.js";
 export * from "./remediation/secret-fix.js";
 export * from "./remediation/bola-report.js";